requirements.txt
torch==2.1.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.1.0+cu118
diffusers==0.24.0
peft==0.3.1
accelerate==0.21.0
transformers==4.30.0
scipy==1.10.1
tqdm==4.65.0
Pillow==9.5.0
